implement Wfsa (#36)

+++ k2/csrc/fsa.h
@@ -116,7 +116,7 @@ struct Fsa {
   int32_t FinalState() const {
     // It's not valid to call this if the FSA is empty.
     CHECK(!arc_indexes.empty());
-    return arc_indexes.size() - 2;
+    return static_cast<int32_t>(arc_indexes.size()) - 2;
   }
 };
